requirements.txt
alpaca-py
flask
git+https://github.com/guyyatsu/ArgumentHandling.git
git+https://github.com/guyyatsu/CryptographyMethods.git
git+https://github.com/guyyatsu/DatabaseAPI.git